+++ src/pages/FaithfulLiturgy.tsx
@@ -1,12 +1,4 @@
-import {
-  Flex,
-  Group,
-  Radio,
-  Select,
-  Skeleton,
-  Text,
-  useMantineTheme,
-} from '@mantine/core';
+import { Flex, Group, Radio, Select, Skeleton } from '@mantine/core';
 import NextButton from '../components/Reusable/NextButton';
 import FormCard from '../components/Reusable/FormCard';
 import PageLayout from '../components/Layout/PageLayout';
@@ -14,12 +6,11 @@ import { useEffect, useState } from 'react';
 import { useNavigate } from 'react-router-dom';
 import axios from 'axios';
 import useDates from '../store/useDates';
-import { useMediaQuery } from '@mantine/hooks';
 import BackButton from '../components/Reusable/BackButton';
 import CardHeader from '../components/Reusable/CardHeader';
 import FormField from '../components/Reusable/FormField';
 import SegControl from '../components/Reusable/SegControl';
-import ReadableDate from '../components/Reusable/ReadableDate';
+import FormHeader from '../components/Reusable/FormHeader';
 
 export interface FaithfulLiturgyApiProps {
   Liturgy3GreatLitanies: string;
@@ -77,8 +68,6 @@ interface FractionObjectProps {
 
 const FaithfulLiturgy = () => {
   const navigate = useNavigate();
-  const theme = useMantineTheme();
-  const isMobile = useMediaQuery(`(max-width: ${theme.breakpoints.sm})`);
 
   const { apiDate, setSelectedCopticDates } = useDates();
   const [fractionObject, setFractionObject] = useState<FractionObjectProps>({
@@ -181,7 +170,7 @@ const FaithfulLiturgy = () => {
   };
 
   const handleSubmit = () => {
-    // Modified Copy of Vespers Data to Post to API
+    // Modified Copy of Faithful Data to Post to API
     const modifiedFaithfulData = { ...faithfulData };
     modifiedFaithfulData.Liturgy3GreatLitanies = faithfulOptions.threeLitanies;
     modifiedFaithfulData.prayerOfReconcilation = [
@@ -216,21 +205,7 @@ const FaithfulLiturgy = () => {
 
   return (
     <PageLayout
-      header={
-        <Flex
-          gap='xl'
-          justify='space-between'
-          align='end'
-          direction='row'
-          wrap='nowrap'
-          sx={{ width: isMobile ? '100%' : '80%' }}
-        >
-          <Text align='left' fw={500}>
-            Selected Date
-          </Text>
-          <ReadableDate />
-        </Flex>
-      }
+      header={<FormHeader />}
       form={
         <FormCard
           content={
